requirements.txt
urllib3==1.26.5
requests==2.25.1
python-dotenv==0.17.1